tests/wpt/tests/css/css-borders/tentative/parsing/box-shadow-color-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'box-shadow-color' with invalid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-color">
<meta name="assert" content="This test checks that 'box-shadow-color' supports only properly defined color values.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_invalid_value("box-shadow-color", "auto");
test_invalid_value("box-shadow-color", "123");
test_invalid_value("box-shadow-color", "#12");
test_invalid_value("box-shadow-color", "#123456789");
test_invalid_value("box-shadow-color", "rgb");
test_invalid_value("box-shadow-color", "rgb(1)");
test_invalid_value("box-shadow-color", "rgb(1,2,3,4,5)");
test_invalid_value("box-shadow-color", "hsla(1,2,3,4,5)");
test_invalid_value("box-shadow-color", "rgb(10%, 20, 30%)");
test_invalid_value("box-shadow-color", "rgba(-2, 300, 400%, -0.5)");
</script>
